backend/requirements.txt
fastapi
uvicorn
faiss-cpu
sentence-transformers
groq
gnews
python-dotenv